requirements.txt
openai>=1.0
python-dotenv>=1.0
mcp[cli]>=1.0
